Avoid duplicate project dirs in playground
This CL fixes a bug in playground where we might end up with duplicate
project directories when Gradle automatically creates a parent project
when it does not exist. When that case happens, Gradle will pick a
project directory relative to the playground root project. Instead, we
provide a project directory relative to the child project to ensure
consistency.

With this change, we can also enable to the project path validation in
the AndroidXPlugin.

Bug: 197253160
Test: cd activity && ./gradlew tasks // fails without this change