Nullify build.gradle of fake parent projects
This change broke navigation because we ended up including
more projects than we used it. It was actually working by
accident because we never set the projectDir on those parent
projects that gradle created. Now we explicitly set them to
non-existing gradle file along with a check in includeProject
to ensure these fakes are not in the build