Adds ability to resolve XConstructorElement into an executable type.
This CL generalizes some types and methods from XMethodElement and
XMethodType so that they also apply to constructors. This is needed
because constructors can also reference type parameters from their
parameter list and we need to be able to resolve those for a particular
type -- similar to XMethodElement#asMemberOf(XType).

To do this, I've moved XMethodElement#executableType() and
XMethodElement#asMemberOf() into XExecutableElement so that both methods
and constructors can get executable types. XExecutableType is a new super
type for both the existing XMethodType and the new XConstructorType.

Test: ./gradlew :room:r-c-p:test
Change-Id: I79b88bdc6a1c5194fbc5d715aaf6481c0e057ef7