Remove actionUpdateContent to enable constructor parameters
Instances of GlanceAppWidget cannot have constructor parameters to enable actionUpdateContent, this constraint causes some bad usability.

Instead remove actionUpdateContent and delegate the responsibility to update the widget to action implementation.

The action might need the GlanceId to update. This requires to bring the GlanceId into the glance module instead of appwidget specific.

Bug: 205716497
Bug: 202208506
Test: ./gradlew :glance:glance:test
Test: ./gradlew :glance:glance-appwidget:test
Test: ./gradlew :glance:glance-appwidget:connectedDevice
Relnote: N/A
Change-Id: I2925a865a6720945ce96430bc5226753560e7900