Validate the API of window extensions.
Validate the API of window extensions to see that the provided methods
are the expected methods. Window extensions ships on device and
therefore might not have the correct interface. If we call a missing
method this would result in a crash app side. We disable window
extensions if any of the expected methods does not match.

Bug: 205630724
Test: manual - load a bad copy of extensions onto a device
Test: ./gradlew window:window:cAT
Change-Id: Ifb39f84ca2501520f68875592fd9183a9c236092