Keyed Modifier.composed variants
Add experimental variants of Modifier.composed that accept keys for
checking the equality of two composed modifiers. Parameters to a
composed modifier factory may be used as keys to produce composed
modifiers that are eligible for skipping optimizations.

Fixes: 205851704
Test: ComposedModifierTest
Relnote: "Add experimental Modifier.composed overloads that accept
keys to compare for equality and qualify for skipping optimizations."

Change-Id: Ice799fbdef230423eab1ad0247979091410b9065