Avoid using `DependencySubstitutions.all`
When a substitution rule is added via `DependencySusbtitutions.all`, Gradle has no
way to determine if that rule will add new Project dependencies or not. As such,
all dependency resolution must be done when building the task execution graph,
in order to determine if the substitution will add additional tasks to the graph.

Using `ResolutionStrategy.eachDependency` avoids this overhead, since the API only permits
one external dependency to be substituted for another, and does not allow Project dependencies
to be added.