Fix order of fields for constructor parameters
This CL works around a KSP bug where it returns the properties for
constructor arguments after regular properties. Both in KAPT and
.class files, they are listed first.

This matters for Room since it sets the order of columns in the
database table.

Note that same problem exists for getters setters but that is not
as important for room so we might tackle it later.

I've also updated the field ordering test to run with sources to
catch these cases.

Bug: 160322705
Test: KspClassFileUtilityTest
Change-Id: I2b3bebbc8a0d4b5556db09599310572ca519b5b9