Add ProGuard rules to keep ambient-related code.
This commit adds ProGuard rules to the androidx.wear library to keep
all the classes and their public methods that are used to implement
ambient support on Wear OS.

We need to keep these because the callbacks are invoked directly by
the platform and ProGuard incorrectly assumes they can be removed,
because no explicit call is present.

Without this, an app that enabled ProGuard will no longer receive the
`onAmbientEnter()` and `onAmbientExit()` callbacks, which means they
are unable to update the app layout to adjust to the new mode.

The :wear:wear-samples-ambient app can be used to validate this by
running `./gradlew :wear:wear-samples-ambient:installRelease` and
getting the device into ambient mode: without this fix, the screen
will still contain the text `Active`, with this fix, the screen will
correctly transition the view to contain the text `Ambient`. This
behavior can also be verified by running `./gradlew
:wear:wear-samples-ambient:installDebug`, which exhibits the correct
behavior independently of whether the ProGuard rules are added since
that sample does not enable ProGuard.

Note that the ProGuard rules are added to the library for both the
debug and release targets, as whether ProGuard is enabled is a choice
of the app using the library rather than the library itself, which can
only provide its ProGuard rules to the target that defines the binary.

Bug: 193775616
Test: ./gradlew :wear:wear-samples-ambient:installRelease
Change-Id: Idaa10fbf067be47a3971b3bbe083c2b5ae7e3fb1