Add support for Rotary Events
Some Wear OS devices contain a physical rotating side button.
When the user turns the button, a rotary input event is
generated and sent to the item in focus. This CL adds two
modifiers, Modifier.onPreRotaryScrollEvent() and
Modifier.onRotaryScrollEvent() that can be used to respond
to rotary events.

Bug: 210748686
Test: ./gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.input.focus.FocusAwareEventPropagationTest
Test: ./gradlew compose:ui:ui:connectedCheck -P android.testInstrumentationRunnerArguments.class=androidx.compose.ui.input.rotary.RotaryScrollEventTest
relnote: "Added Modifier.onRotaryScrollEvent() and Modifier.onPreRotaryScrollEvent() for wear devices with a rotating side button"
Change-Id: I18bf59262881d9c4a2edea159751164340ac3ed5