Correct ScalingLazyListItemInfo.offset for scaled list items
There is an error in calculation of the offset in ScalingLazyListItemInfo for items which are scaled which is resulting in incorrect calculation of PositionIndicator ScalingLazyColumnStateAdapter results. This results in list items which are actually visible being reported as being outside of the visible viewport range. This in turn results in incorrect calculations in ScalingLazyColumnStateAdapter which results in the position indicator appearing to grow and shrink in size as the users scrolls content.

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 214056513
Change-Id: Iaf6b13c8acb79588032ad37eeaa6190d4dc96edd