Add zero-arg persistableBundleOf() overload
This saves the compiler from needing to allocate an empty array for the
vararg parameter when invoking persistableBundleOf() without arguments.

Developers will expect persistableBundleOf() to work like listOf() and
return an empty PersistableBundle and are likely to idiomatically perfer
`persistableBundleOf()` over `PersistableBunde(0)` so I expect this path
to get enough usage to justify the additional method.

Test: PersistableBundleTest
Relnote: "Added zero-arg overload for persistableBundleOf() to avoid an
    unnecessary array allocation when creating an empty
    PersistableBundle"
Bug: 214270424
Change-Id: Icd7a41c8211462f6c0f2b9554c30c1fcfe8a1c16