Fix Robolectric dependencies in MPP Compose
When defining dependencies with multi platform enabled, we need to use
`androidTest` as the name of the local test source set because it is
multi platform. However, AGP also adds these dependencies to the
instrumented test source set. This is a problem for example when you
include Robolectric in your local test dependencies.

To solve this, declare dependencies that must be kept out of
instrumented tests in a `dependency` block for AGP only.

Bug: 214407011
Test: Ran local tests of all affected modules with multi-platform
enabled

Change-Id: I7e87509d53beb31875d956506d5a57222a860244