Fix problem with multiple JankStats per window
Previously, creating multiple instances of JankStats on a particular
window would potentially create multiple OnFrameMetricsAvailableListener
objects on that window, causing both unnecessary overhead as well as
an issue where one Jankstats instance exclusively consumes data that
should be passed to multiple listeners (SingleFrameStates).

This fix creates only one listener per window, along with a list
of listeners that it delegates to. Creating a new JankStats instance
will only create/add the core window listener if none yet exists.
Otherwise, it will simply add its listener to the delegate list, which
will be called by the single listener when it is called by Window.

A similar change was made to the JankStatsApi16Impl to have a single
OnPreDrawListener which delegates to multiple instances (if they exist).

Test: ran manual integration tests and added new automated test
Bug: 213499234
Change-Id: I72b7e703bdd476b8bd8fc32f73d6932b9086478a