Use method name when partitioning for overrides
Kotlin methods / accessors can be annotated with JvmName to change their
name in the byte code. This is designed to be invisible from kotlin
code. As such, it is also not allowed on overriding methods or open
methods. Unfortunately, it is also possible for the developer to
suppress that error via @Suppress("INAPPLICABLE_JVM_NAME").

This CL updates XProcessing to basically ignore that annotation for
overrides. The practical impact of this is a 30x performance gain for
getAllMethods since we don't need to resolve method types to check for
overrides.

Upon discussions with the Kotlin Compiler team, their feedback is that
it is fair to ignore that annotation since the behavior of the compiler
gets really complicated when used along with method overrides.
Furthermore, user suppressed an error hence they cannot really expect
everything to work as usual.

I've added a warning for Dao and Database methods that checks if the
method is annotated with JvmName, and if so, prints a warning that
cannot be suppressed. Also added tests for cases where it happens to
work so that we know if we further break it :) .

Measurements: (when it was a config, but should be relatively the same)

androidx.room.compiler.processing.profiling.GetMethodsScenarioTest > allMethods_filterOutInline STANDARD_OUT
    Stats(min=69.319659ms, max=145.664166ms, avg=91.519145ms, mean=70.507303ms, allMeasurements=[88.612249ms, 78.264850ms, 88.811043ms, 70.880497ms, 71.694948ms, 69.319659ms, 145.664166ms, 90.055662ms, 130.070210ms, 81.818174ms])

androidx.room.compiler.processing.profiling.GetMethodsScenarioTest > allMethods_filterOutInline PASSED

androidx.room.compiler.processing.profiling.GetMethodsScenarioTest > allMethods_keepInline STANDARD_OUT
    Stats(min=1.643961ms, max=3.542199ms, avg=2.149422ms, mean=2.636825ms, allMeasurements=[2.081342ms, 1.933364ms, 1.709868ms, 1.679751ms, 3.542199ms, 1.731452ms, 1.643961ms, 2.329103ms, 1.686483ms, 3.156701ms])

Bug: 160322705
Test: GetMethodsSecenarioTest, XTypeElementTest, DaoProcessorTest,
Test: DatabaseProcessorTest, JvmNameInDaoTest.kt

Merged-In: I0e6415e5a17f378f7009adff53f1c9e65930fd7d
Change-Id: I0e6415e5a17f378f7009adff53f1c9e65930fd7d