ScalingLazyColumn.kt - reduce amount of extra space allocated to allow additional list items to be composed.
Collection of performance improvements for the ScalingLazyColumn. The primary change is to reduce the amount of space allowed of extra list items that are composed in the underlying LazyList from 20% to 10%. The extra list items are composed in case they are needed as a result of the scaling down in size of other list components. However for the standard list scaling we don't need an additional 20% both above and below the currently visible items - 10% should be more than sufficient. This behaviour can be overridden in the API if additional items are needed for a given scaling strategy.

Other minor performance enhancements have been added as the result of profiling with perfetto.

Test: ./gradlew :wear:compose:compose-material:connectedCheck --info --daemon
Bug: 214993637
Change-Id: I4dbc988c0436362ed6dee963105d168b35ac350e