Fix jvm setter name calculation
This CL fixes a bug in kotlin metadata where, for properties that start
with is but do not continue with a capital letter, we would apply the
`is` rule while generating the getter/setter but that is not how kotlin
compiler works.

This time, I copied implementations from the kotlin compiler codebase to
ensure they match what kotlin does.

Bug: 215069358
Test: PojoProcessorTest

Change-Id: I8baa4b048e102e2b6390f05e2fab8140981f962d