Fix improper destination nesting with deep links
When deeplinking to a destination that has multiple nested nav_graphs,
the handleDeepLink() function will incorrectly remove the start
destinations of the parent NavGraphs when navigating to child graphs of
the parent.

What we really want is that you only pop and save destinations when
navigating to sibling graphs, not child graphs.

RelNote: "Intermediate start destinations will now be present when
deep linking through multiple nested `NavGraph`s."
Bug: 214383060
Test: tested in sample app in bug
Test: verified no regressions for other handleDeepLink cases

Change-Id: I504c04d2cc4381af22405266192ea0f5094f9c16