Export and read in atomic library group data
This is a re-implementation of aosp/1907951 and aosp/1954617, but with a different method of storing and retrieving the atomic library group list.

The lint-checks project will now export the atomic library group data before running any lint tasks, and package the file in lint-checks.jar.  The BanInappropriateExperimentalUsage detector will then read in this file as needed to check if a library's group is atomic.

This change also removes the export task from AndroidXImplPlugin.

Bug: 204371669

Test: bOS
Test: ./gradlew :lint-checks:test -Pandroidx.verifyUpToDate

The following tests should no longer be necessary, but are included for posterity as they surfaced up issues in the previous implementation.  For even more coverage, also run the following with `Debug` replaced with `Release`.

./gradlew --rerun-tasks --no-build-cache --no-daemon :compose:runtime:runtime:lintDebug

./gradlew --rerun-tasks --no-build-cache --no-daemon :room:integration-tests:room-testapp-kotlin:lintWithKspDebug

Change-Id: Ie07754777a8a6494f0070da9c9930209f4dcc8f0