Remove lifecycle lint baseline suppressions
Although we have taken the proper precausions for making newer API calls, lint may not be able to tell. Instead of suppressing the warning using lint baseline, we will create static classes with the appropriate API level for the methods we need.

Suppressed lint for mutable liveData. Also added RequireApi annotations in test.

Test: ./gradlew bOS
Fixes: 188452852
Change-Id: I471cd0ed4e829b8f878fff25e379f7f1a31d5b73