Use latest ndk for presubmit and remove inspection from playground
This CL updates the presubmit yml to set the ndk dir to the latest.
We should eventually followup with a custom task that sets it to
whatever androidx expects.

I've also removed inspection from the github build since it seems to be
referencing internal prebuilts. To allow that, I had to change the
inspection plugin to accept null inspector projects (only in
playground).

Also increased the gradle memory limit to 6g.

Bug: 216535050
Test: CI