Use exact version of NDK needed by presubmit in Github Actions
It is required to install ninja manually due to it not being included in
the virtual env provided by Github Actions runner.

See https://github.com/actions/virtual-environments/issues/741

In order for Ninja to be visible to CMake during NDK builds, it must be
installed in the same subdirectory, so for now it is symlinked on Linux
builds. See b/206099937

Fixes: 216535050
Test: Github CI
Change-Id: I1a741b1765822f0004adc3985b7babc8022046a5