Don't save Gradle User Home for branch workflows
GitHub cache entries are not shared between builds on different branches.
This means that identical cache entries will be stored separately for different branches.
The exception to the is cache entries for the default (androidx-main) branch can be read by
actions invoked for other branches.

To avoid having private branch-specific entries evict shared entries,
we configured branch workflows to use the cache read-only.