Pin wear.compose.navigation to navigation 2.4.0
The navigation libraries are stable on 2.4.0 and will make point
releases for bug fixes as required. Wear does not depend on the
upcoming 2.5.x features, so better not to depend on 2.5, in case that
gets stuck in alpha due to changes down their dependency tree.
Future navigation libraries will be backwards compatible for all
earlier versions.

Bug: 217191557

Test: N/A
Change-Id: I0d77754bd362549e29473feec96a89538d8c7b99