Remove unused local build-cache entries on CI
By default, unused local build-cache entries are retained for 7 days.
In the context of a GitHub Actions workflow:
- these old cache entries take up additional space in the GitHub Actions cache
- these old cache entries are very unlikely to be used again at a later date
As such, there is no benefit to keeping them around in the restored Gradle User Home.

This change reduces the time these entries are retained to the minimum, a single day.
Cache cleanup will be performed every 24hrs, and at that time all build-cache entries
that have not been used in the previous day will be deleted.