Use sub action to build individual projects
This CL refactors the workflow to use a sub action to build individual
projects. This helps cleanup the build and also have an easy way to
disable certain builds.

I've also added a new ci-config.json file to control which projects are
included / excluded along with a companion script to decide based on the
current branch.

For now, we have two setups, 1 for compose and 1 for the rest. Compose
branch only enables compose projects meanwhile the main branch enables
everything by default.

Bug: n/a
Test: CI