Ensure SavedStateHandle actually sets defaults
When calling getLiveData() on SavedStateHandle with an initial value, it
never stored the initial value so consequent calls to get() with the
same key would still return null.

We should make sure we actually store the value with its key.

RelNote: "`SavedStateHandle` now properly stores the defaultValue when
no value for the specified key exists."
Test: SavedStateHandleTest
Bug: 178510877

Change-Id: I1c6ce7584520393eaf58884f92e0b6d90ff7de12