Use programmatic configuration for build-cache cleanup
A previous attempt to do this via a CLI argument failed, since the argument did
not apply to the cache settings for the `playground-plugin` included build.
Since the `playground-plugin` settings are loaded first, the CLI settings were
being ignored.

This change configures the cleanup threshold directly in Gradle settings, and does
so both for the Playground included build and the main build. This results in
any build-cache entries that have not been used in the previous day to be removed.