Log instead of throwing if SQLite process lock can't be created.
It might be the case that the lock file can't be created due to a bad database path or a path where the app has no permission. In such cases just don't grab a file lock, log instead of throwing as it might be very likely the database might fail to open.

Bug: 217738355
Test: n/a
Change-Id: I73bb02863795d63bee6b7824174a245715d0b310