Add namespace support in safeArgs
We currently retrieve the package from the namespace for
ApplicationVariants, but we fail to apply it to the R resources when a
package name is missing from the manifest.

We should support using namespace when the package is missing from the
manifest to ensure we can retrieve xml resources from the proper
location.

If you would like to alternative packages for different product flavors
you still must use the package in the manifest.

RelNote: "Navigation SafeArgs now supports using the namespace defined
in the `build.gradle` in place of the package in the AndroidManifest."
Test: changed the kotlin test to use namespace instead of package
Bug: 217414933

Change-Id: I659efc0f5b74619f5fe3ffb882c96ccb360670b7