Add caption3 to the Wear Material Theme Typology
At the request of our UX team, and to bring Wear Compose Material into line with Tiles we have added a new extra small font size (caption3). We have also updated the KDoc descriptions of the other members of the Typography in line with latest descriptions from the UX team.

Bug: 220128356
RelNote: "We have added a new entry caption3 to the Wear Material Theme Typology.Caption3 is an small font used for the extra long-form writing like legal texts."
Test: ./gradlew wear:compose:compose-material:connectedCheck --info --daemon
Change-Id: I74b133e3dbb660adf576893cfedf7bec95042c9b