Hide top-level dialog from the taskbar
Swing has a special constructor for that (without arguments). It says:
```
A shared, hidden frame will be set as the owner of the dialog.
```

This will hide the icon in the taskbar.
We need exactly this as a default behaviour for any Dialog, even if it is top-level.
If the user need to show the icon in the taskbar, Window is a better option.

The reproducer:
```
import androidx.compose.ui.awt.ComposeDialog
import javax.swing.SwingUtilities

fun main() = SwingUtilities.invokeLater {
    val dialog = ComposeDialog()
    dialog.setSize(800, 600)
    dialog.isVisible = true
}
```

The icon shouldn't be in the taskbar now.

Test: manual. See the snippet
Change-Id: I537a46948206e90dfcfdd2849d45764067fb29b4