Optimize getWatchFaceOverlayStyle
Previously getWatchFaceOverlayStyle was waiting for
deferredWatchFaceImpl and was dispatched on the busy UI thread.
It only needs to wait for the WatchFace (not WatchFaceImpl) to be
constructed and it can run enturely on a binder thread. Perfetto
traces with the androidx sample watch faces have show a latency
reduction from ~100 ms to ~1ms.

Bug: 218991776
Test: Manual testing with perfetto
Change-Id: Ia7301ec3d1ce0d05b10cc64af8baae0d6d39a9ad